:bug: Make vertex normals scale-safe and reproducible
Closes both of plans/review.md's standing open items. Neither was in the four
files the review pass set out to read; both were found by the probe written for
solve_normal_equations' scale dependence in edec5f7, which is the argument for
sweeping the scale axis whenever a numerical finding is in hand.

A mesh below scale ~3e-6 had no vertex normals at all. face_normals_and_area
returned the raw cross product -- magnitude 2 * area, where a *unit* normal was
promised -- whenever |cross| <= TOLERANCE_ZERO, and |cross| scales as h^2, so at
h ~ 3e-6 every face of the mesh took that branch. vertex_normals then
area-weights those, squaring the smallness, and wp.normalize sees a float32
length_sq underflowed to exactly zero: measured, every vertex normal on
icosphere(3) came back zero at h = 3e-6 and below, and with them every
curvature, with nothing raised.

The fix is to delete the tolerance rather than make it relative. Dividing by
norm is well conditioned for any positive norm, and at norm == 0 the cross
product *is* the zero vector, so the two branches already agreed there and the
floor was never protecting the division. That also makes the function agree with
its sibling face_normal / triangle_normal, which is section 2.4's
one-predicate-one-spelling rule. The only faces whose answer changes are those
that used to get a non-unit "unit normal", and the direction is unchanged at
every one, so no consumer that reads a sign (intersection's plane side) or an
angle (remesh's feature test, through the scale-covariant vector_angle) can see
it. Normals are unit down to h = 1e-9 now, where float32 storage of the
positions is the next limit; principal_curvature is scale-equivariant across
1e-9 to 1e+6. The road not taken: a relative test against the face's own edge
lengths costs two square roots per face and a convention change at 74
TOLERANCE_ZERO sites across 14 files.

vertex_normals was also not reproducible run to run on CUDA, and the cheap fix
turned out to be the accumulator's *width*, not its summation order -- which is
what the open item lacked. It asked whether a deterministic rebuild (a segmented
reduction over a vertex-face CSR) was worth the build every caller would pay; it
is not, because widening the (n, 3) accumulator to float64 reaches the same
place for free. Nothing fixes the ordering -- Warp exposes no deterministic
float atomic -- but at float64 two orderings disagree by ~1e-16 relative, far
below what the returned float32 can represent, so the answer is reproducible in
practice: bit-identical over eight runs where the float32 accumulator moved
every time. Not a guarantee, and atomic_add_vec3's comment says so. Cost against
a detached baseline worktree: 1.29x faster at 10k vertices, level at 164k, 1.10x
slower at 655k -- the win is the launch the fused narrow-and-normalize tail
removes (a float32 accumulator could reach wp.vec3 through a zero-copy
array_cast; a float64 one cannot), the loss is the wider atomics once the call
stops being launch-bound.

The open item's own account of the symptom was wrong, and so was the remedy
principal_curvature advertised. It read as an umbilic swap, the pair returned
transposed where the ordering carries no information. Measured: the *sorted*
pair moves by the same 7.96e-04, so the values themselves move, at 19 of 544
vertices, worst 77% relative, and they are the near-*flat* vertices where the
quadric fit is ill conditioned rather than the umbilic ones. And the Notes told
a caller who needed reproducibility to pass its own face_normals and face_areas
-- measured to change the spread by exactly nothing, since the non-determinism
is in the vertex-normal scatter, which runs either way. A remedy stated in a
docstring is a claim to run once before believing. The Notes now carry the
conditioning fact, which is the part that survived and the part a caller can act
on.

Four regression tests, mutation-probed against a detached pre-fix worktree: the
two reproducibility ones assert exact equality deliberately, since the old
behaviour fails that and a 1e-6 tolerance would not.

Gates: full suite 3347 passed / 1 skipped, ruff clean, basedpyright 0 errors,
mkdocs --strict clean, tests.parity 584 pairs unchanged,
benchmarks/test_meshes.py 80 passed.

+++ triwarp/kernels/scatter.py
@@ -14,11 +14,31 @@ from triwarp.kernels.grouping import sorted_run_start
 
 
 @wp.func
-def atomic_add_vec3(out_sum: wp.array2d[wp.float32], row: wp.int32, v: wp.vec3) -> None:
+def atomic_add_vec3(out_sum: wp.array2d[wp.float64], row: wp.int32, v: wp.vec3) -> None:
     # Component-wise atomic accumulation of a wp.vec3 into row ``row`` of a (n, 3) buffer.
-    wp.atomic_add(out_sum, row, 0, v[0])
-    wp.atomic_add(out_sum, row, 1, v[1])
-    wp.atomic_add(out_sum, row, 2, v[2])
+    #
+    # **The accumulator is float64 while the values are float32, and that is what makes the sum
+    # reproducible.** A float atomic's summation order is whatever the scheduler hands it, and
+    # float addition is not associative, so a float32 accumulator's answer moves by about one ULP
+    # (measured 1.19e-07 on vertex normals) between runs of the identical launch. That is
+    # harmless in itself and is not harmless downstream: ``curvature.principal_curvature`` fits an
+    # ill-conditioned quadric to these normals and turns it into swings of up to **77 %** of the
+    # returned curvature at near-flat vertices. Widening the accumulator does not fix the ordering
+    # -- nothing here can -- but it drops the disagreement between two orderings to ~1e-16
+    # relative, far below what the float32 result can represent, so the narrowed answer is
+    # reproducible in practice (measured: bit-identical over eight runs where the float32
+    # accumulator moved every time). It is not a *guarantee*: a sum landing within 1e-16 of a
+    # float32 rounding boundary could still round both ways.
+    #
+    # Cost, measured on an RTX 5090 against a detached baseline worktree: the scatter kernel alone
+    # is 0.99-1.22x the float32 one from 82k to 1.3M faces, and ``vertices.vertex_normals`` end to
+    # end is **1.29x faster** at 10k vertices, level at 164k and **1.10x slower** at 655k -- the
+    # win being the launch the fused narrow-and-normalize tail removes (a float32 accumulator could
+    # reach ``wp.vec3`` through a zero-copy ``array_cast``, a float64 one cannot), and the loss
+    # being the wider atomics once the call stops being launch-bound. The (n, 3) buffer doubles.
+    wp.atomic_add(out_sum, row, 0, wp.float64(v[0]))
+    wp.atomic_add(out_sum, row, 1, wp.float64(v[1]))
+    wp.atomic_add(out_sum, row, 2, wp.float64(v[2]))
 
 
 @wp.kernel
@@ -82,7 +102,7 @@ def scatter_sum_scalar(
 
 @wp.kernel
 def scatter_sum_vec(
-    values: wp.array[wp.vec3], indices: wp.array2d[wp.int32], out_sum: wp.array2d[wp.float32]
+    values: wp.array[wp.vec3], indices: wp.array2d[wp.int32], out_sum: wp.array2d[wp.float64]
 ) -> None:
     tid = wp.int32(wp.tid())
     index = indices[tid]
@@ -96,7 +116,7 @@ def scatter_weighted_sum_vec(
     values: wp.array[wp.vec3],
     indices: wp.array2d[wp.int32],
     weights: wp.array2d[wp.float32],
-    out_sum: wp.array2d[wp.float32],
+    out_sum: wp.array2d[wp.float64],
 ) -> None:
     tid = wp.int32(wp.tid())
     index = indices[tid]
